Resources page
@mark-bradshaw I need to get with you and see how to wire up pages.
I’ve added the resources page and routes here but fear I may be off
since we will need it to be dynamic. I’ve added the page in the
database but now need to hook it up with keystone. Let me know if you
might have some time tonight to look into it with me.